backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gigachat==0.1.43
requests==2.31.0
python-pptx==1.0.2
Pillow==11.0.0
gunicorn==21.2.0
whitenoise==6.6.0
